+++ K8sUtils/public/Get-PodEvent.ps1
@@ -33,17 +33,32 @@ function Get-PodEvent {
         [string] $Namespace = "default"
     )
     Write-Verbose "kubectl get events --namespace $Namespace --field-selector `"involvedObject.name=$PodName`" -o json"
-    $events = kubectl get events --namespace $Namespace --field-selector "involvedObject.name=$PodName" -o json | ConvertFrom-Json
+    $json = kubectl get events --namespace $Namespace --field-selector "involvedObject.name=$PodName" -o json
 
+    Write-Verbose "kubectl exit code: $LASTEXITCODE"
     if ($LASTEXITCODE -ne 0) {
+        Write-Status "kubectl get events --namespace $Namespace --field-selector `"involvedObject.name=$PodName`" -o json" -LogLevel warning
+        Write-Status "  had exit code of $LASTEXITCODE" -LogLevel warning
+        Write-Status "  JSON is $json" -LogLevel warning
         return $null
     }
+    $events = $json | ConvertFrom-Json
     if ($null -eq $events) { # valid if no events, such as it didn't need to create a new pod
-        return @()
+        Write-Verbose "Null events after conversion"
+        Write-Output @() -NoEnumerate # Prevent @() from turning into $null
+        return
+    }
+    if ($events.items) {
+        Write-Verbose "Events count: $($events.items.count)"
     }
     if ($NoNormal) {
-        $events.items | Where-Object { $_.type -ne "Normal" }
+        $ret = $events.items | Where-Object { $_.type -ne "Normal" }
     } else {
-        $events.items
+        $ret = $events.items
     }
+    if ($null -eq $ret) {
+        Write-Verbose "Null items, returning empty array"
+        $ret = @()
+    }
+    Write-Output $ret -NoEnumerate
 }